requirements.txt
python-telegram-bot==20.7
pillow
qrcode[pil]
python-dotenv
